Moved PDF generation to RefusjonRoutes
